+++ report/data_gen/src/colfind/query_evenness.rs
@@ -52,7 +52,7 @@ pub fn handle2(fb: &mut FigureBuilder, prefix: &str, grow: f64, num_bots: usize)
                 .vistr()
                 .dfs_inorder_iter()
                 .enumerate()
-                .map(|(i, element)| [i as f64, *element as f64])
+                .map(|(i, element)| [i as f32, *element as f32])
         );
 
         fb.finish_plot(splot, move_format!("query_evenness_theory_{}", prefix));
@@ -77,7 +77,7 @@ pub fn handle2(fb: &mut FigureBuilder, prefix: &str, grow: f64, num_bots: usize)
         tree.vistr()
             .dfs_inorder_iter()
             .enumerate()
-            .map(|(i, element)| [i as f64, element.range.len() as f64])
+            .map(|(i, element)| [i as f32, element.range.len() as f32])
     );
 
     fb.finish_plot(splot, move_format!("query_num_per_node_theory_{}", prefix));
