+++ report/data_gen/src/spiral.rs
@@ -36,7 +36,7 @@ fn handle_num(fb: &mut FigureBuilder) {
 
     plot.line(
         "",
-        rects.iter().map(|x| [x.0 as f64, x.1 as f64]),
+        rects.iter().map(|x| [x.0 as f32, x.1 as f32]),
     );
 
     fb.finish_plot(plot, "spiral_data_num");
@@ -69,7 +69,7 @@ fn handle_grow(fb: &mut FigureBuilder) {
 
     plot.line(
         "",
-        rects.iter().map(|x| [x.0 as f64, x.1 as f64]),
+        rects.iter().map(|x| [x.0 as f32, x.1 as f32]),
     );
 
     fb.finish_plot(plot, "spiral_data_grow");
@@ -92,7 +92,7 @@ fn handle_visualize(fb: &mut FigureBuilder) {
         "",
         make(DEFAULT_GROW)
             .into_iter()
-            .map(|v| [v.x as f64, v.y as f64])
+            .map(|v| [v.x as f32, v.y as f32])
     );
 
     fb.finish_plot(plot, "spiral_visualize");
